feat: add delete loan validator.

+++ Bibliotrack.Application/Commands/LoanCommands/DeleteLoan/DeleteLoanHandler.cs
@@ -7,10 +7,12 @@ namespace Bibliotrack.Application.Commands.LoanCommands.DeleteLoan
     public class DeleteLoanHandler : IRequestHandler<DeleteLoanCommand, ResultViewModel>
     {
         private readonly ILoanRepository _loanRepository;
+        private readonly IBookRepository _bookRepository;
 
-        public DeleteLoanHandler(ILoanRepository loanRepository)
+        public DeleteLoanHandler(ILoanRepository loanRepository, IBookRepository bookRepository)
         {
             _loanRepository = loanRepository;
+            _bookRepository = bookRepository;
         }
 
         public async Task<ResultViewModel> Handle(DeleteLoanCommand request, CancellationToken cancellationToken)
@@ -20,6 +22,12 @@ namespace Bibliotrack.Application.Commands.LoanCommands.DeleteLoan
             if (loan == null)
                 return ResultViewModel.Error("Loan not found.");
 
+            if (loan.ReturnDate != null)
+                return ResultViewModel.Error("Cannot delete a returned loan.");
+
+            loan.Book.IncreaseQuantity();
+
+            await _bookRepository.Update(loan.Book);
             await _loanRepository.Delete(loan.Id);
 
             return ResultViewModel.Success();
